+++ CHANGELOG.md
@@ -91,6 +91,7 @@ All notable changes to this project will be documented in this file.
   - Re-enabled the filament manager spool assignment:
     - Without filament manager plugin this will be a multi-select option. You can CTRL + Click to select mutliple, or normal left click to select a single. 
     - With filament manager plugin this will be a single dropdown menu only allowing to select a single spool as per the plugins requirements. 
+  - Printers with selected spools are now disabled from selection without filament manager plugin.
 
 ### Removed